requirements.txt
# Backend dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Document ingestion
pypdf2>=3.0.0

# Embeddings and vector store
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4

# HTTP for Ollama
httpx>=0.26.0
sse-starlette>=1.8.0
